Q2/requirements.txt
tika==1.24
tabula-py==2.3.0
pyPDF2==1.26.0
numpy==1.19.5